conditional added

+++ myce_masternode_setup.sh
@@ -15,36 +15,41 @@ echo "**********************************************************************"
 echo && echo && echo
 
 
-sudo apt-get update
-sudo apt-get -y upgrade
-sudo apt-get install -y software-properties-common
-sudo add-apt-repository -y ppa:bitcoin/bitcoin
-sudo apt-get update
-sudo apt-get install -y build-essential libtool autotools-dev automake pkg-config libssl-dev libevent-dev bsdmainutils libboost-all-dev libdb4.8-dev libdb4.8++-dev libminiupnpc-dev libzmq3-dev tmux
-sudo apt-get install -y libgmp3-dev
+echo "Would you like to install your beautiful brand new MYCE masternode ? [y/n]"
+read DOSETUP
 
-sudo apt-get update
+if [[ $DOSETUP =~ "y" ]] ; then
+  sudo apt-get update
+  sudo apt-get -y upgrade
+  sudo apt-get install -y software-properties-common
+  sudo add-apt-repository -y ppa:bitcoin/bitcoin
+  sudo apt-get update
+  sudo apt-get install -y build-essential libtool autotools-dev automake pkg-config libssl-dev libevent-dev bsdmainutils libboost-all-dev libdb4.8-dev libdb4.8++-dev libminiupnpc-dev libzmq3-dev tmux
+  sudo apt-get install -y libgmp3-dev
 
-cd /var
-sudo touch swap.img
-sudo chmod 600 swap.img
-sudo dd if=/dev/zero of=/var/swap.img bs=1024k count=2000
-sudo mkswap /var/swap.img
-sudo swapon /var/swap.img
-sudo free
-sudo echo "/var/swap.img none swap sw 0 0" >> /etc/fstab
-cd
+  sudo apt-get update
 
-sudo apt-get install -y ufw
-sudo ufw allow ssh/tcp
-sudo ufw logging on
-#sudo ufw enable
-sudo ufw status
+  cd /var
+  sudo touch swap.img
+  sudo chmod 600 swap.img
+  sudo dd if=/dev/zero of=/var/swap.img bs=1024k count=2000
+  sudo mkswap /var/swap.img
+  sudo swapon /var/swap.img
+  sudo free
+  sudo echo "/var/swap.img none swap sw 0 0" >> /etc/fstab
+  cd
 
-mkdir -p ~/bin
-echo 'export PATH=~/bin:$PATH' > ~/.bash_aliases
-source ~/.bashrc
+  sudo apt-get install -y ufw
+  sudo ufw allow ssh/tcp
+  sudo ufw logging on
+  #sudo ufw enable
+  sudo ufw status
 
+  mkdir -p ~/bin
+  echo 'export PATH=~/bin:$PATH' > ~/.bash_aliases
+  source ~/.bashrc
+
+fi
 
 git clone https://github.com/mycelliumcoin/MycelliumMN
 cd MycelliumMN/src/leveldb && chmod 777 * && cd .. && make -f makefile.unix
@@ -84,8 +89,8 @@ echo "**********************************************************************"
 echo ""
 
 
-CONF_DIR=~/.myce/
-CONF_FILE=myce.conf
+CONF_DIR=~/.Myce/
+CONF_FILE=Myce.conf
 
 echo "rpcuser=rpccheese"`shuf -i 100000-10000000 -n 1` >> $CONF_DIR/$CONF_FILE
 echo "rpcpassword=rpccheesepass"`shuf -i 100000-10000000 -n 1` >> $CONF_DIR/$CONF_FILE
